+++ src/maple_be/src/be/switch_lowerer.cpp
@@ -59,7 +59,8 @@ void SwitchLowerer::FindClusters(MapleVector<Cluster> &clusters) {
   while (i < length - kClusterSwitchCutoff) {
     for (int32 j = length - 1; j > i; --j) {
       float tmp1 = static_cast<float>(j - i);
-      float tmp2 = static_cast<float>(stmt->GetCasePair(j).first) - static_cast<float>(stmt->GetCasePair(i).first);
+      float tmp2 = static_cast<float>(stmt->GetCasePair(static_cast<size_t>(static_cast<uint>(j))).first) -
+          static_cast<float>(stmt->GetCasePair(static_cast<size_t>(static_cast<uint>(i))).first);
       if (((j - i) >= kClusterSwitchCutoff) &&
           (tmp2 < kMaxRangeGotoTableSize) &&
           ((tmp1 / tmp2) >= kClusterSwitchDensity)) {
